optimize_1q_gates pass (#1442)
* optimize_1q_gates moved as a pass

* new pass file and tests

* style

* lint

* style

* import order

* redundant import

* bad import

* style

* move out requires

* docstring

* running unrolling before optimization

* style

* lint

* unifying FakeBackend

* a test using transpiler

* style

* cx is unrolled to CX

* Update qiskit/transpiler/passes/optimize_1q_gates.py


* a specific basis is needed before optimize

* style

* cx

* append
