+++ qiskit/mapper/__init__.py
@@ -13,5 +13,5 @@ from ._couplingerror import CouplingError
 from ._layout import Layout
 from ._mappererror import MapperError
 from ._mapping import (swap_mapper, direction_mapper, cx_cancellation,
-                       optimize_1q_gates, remove_last_measurements,
+                       remove_last_measurements,
                        return_last_measurements)
