+++ qiskit/mapper/_mapping.py
@@ -14,25 +14,18 @@ import logging
 import pprint
 import sys
 
-import networkx as nx
 import numpy as np
 import sympy
-from sympy import Number as N
 
 from qiskit.qasm import _node as node
 from qiskit.mapper import MapperError
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.dagcircuit._dagcircuiterror import DAGCircuitError
-from qiskit.quantum_info.operators.quaternion import quaternion_from_euler
 from qiskit import QuantumRegister
 from qiskit.extensions.standard.h import HGate
 from qiskit.extensions.standard.cx import CnotGate
 from qiskit.extensions.standard.swap import SwapGate
-from qiskit.extensions.standard.u1 import U1Gate
-from qiskit.extensions.standard.u2 import U2Gate
-from qiskit.extensions.standard.u3 import U3Gate
 from qiskit.circuit.measure import Measure
-from qiskit.circuit.instruction import Instruction
 
 logger = logging.getLogger(__name__)
 
@@ -607,53 +600,6 @@ def swap_mapper(circuit_graph, coupling_graph,
     return dagcircuit_output, initial_layout, last_layout
 
 
-def yzy_to_zyz(xi, theta1, theta2, eps=1e-9):
-    """Express a Y.Z.Y single qubit gate as a Z.Y.Z gate.
-
-    Solve the equation
-
-    .. math::
-
-    Ry(theta1).Rz(xi).Ry(theta2) = Rz(phi).Ry(theta).Rz(lambda)
-
-    for theta, phi, and lambda.
-
-    Return a solution theta, phi, and lambda.
-    """
-    Q = quaternion_from_euler([theta1, xi, theta2], 'yzy')
-    euler = Q.to_zyz()
-    P = quaternion_from_euler(euler, 'zyz')
-    # output order different than rotation order
-    out_angles = (euler[1], euler[0], euler[2])
-    abs_inner = abs(P.data.dot(Q.data))
-    if not np.allclose(abs_inner, 1, eps):
-        logger.debug("xi=%s", xi)
-        logger.debug("theta1=%s", theta1)
-        logger.debug("theta2=%s", theta2)
-        logger.debug("solutions=%s", out_angles)
-        logger.debug("abs_inner=%s", abs_inner)
-        raise MapperError('YZY and ZYZ angles do not give same rotation matrix.')
-    return out_angles
-
-
-def compose_u3(theta1, phi1, lambda1, theta2, phi2, lambda2):
-    """Return a triple theta, phi, lambda for the product.
-
-    u3(theta, phi, lambda)
-       = u3(theta1, phi1, lambda1).u3(theta2, phi2, lambda2)
-       = Rz(phi1).Ry(theta1).Rz(lambda1+phi2).Ry(theta2).Rz(lambda2)
-       = Rz(phi1).Rz(phi').Ry(theta').Rz(lambda').Rz(lambda2)
-       = u3(theta', phi1 + phi', lambda2 + lambda')
-
-    Return theta, phi, lambda.
-    """
-    # Careful with the factor of two in yzy_to_zyz
-    thetap, phip, lambdap = yzy_to_zyz((lambda1 + phi2),
-                                       theta1, theta2)
-    (theta, phi, lamb) = (thetap, phi1 + phip, lambda2 + lambdap)
-    return (theta, phi, lamb)
-
-
 def cx_cancellation(circuit):
     """Cancel back-to-back "cx" gates in circuit."""
     runs = circuit.collect_runs(["cx"])
@@ -679,162 +625,6 @@ def cx_cancellation(circuit):
                 for n in chunk[1:]:
                     circuit._remove_op_node(n)
     return circuit
-
-
-def optimize_1q_gates(circuit):
-    """Simplify runs of single qubit gates in the QX basis.
-
-    Return a new circuit that has been optimized.
-    """
-    from qiskit.transpiler.passes.mapping.unroller import Unroller
-    qx_basis = ["u1", "u2", "u3", "cx", "id"]
-    unrolled = Unroller(qx_basis).run(circuit)
-
-    runs = unrolled.collect_runs(["u1", "u2", "u3", "id"])
-    for run in runs:
-        run_qarg = unrolled.multi_graph.node[run[0]]["qargs"][0]
-        right_name = "u1"
-        right_parameters = (N(0), N(0), N(0))  # (theta, phi, lambda)
-        for current_node in run:
-            nd = unrolled.multi_graph.node[current_node]
-            left_name = nd["name"]
-            if (nd["condition"] is not None
-                    or len(nd["qargs"]) != 1
-                    or nd["qargs"][0] != run_qarg
-                    or left_name not in ["u1", "u2", "u3", "id"]):
-                raise MapperError("internal error")
-            if left_name == "u1":
-                left_parameters = (N(0), N(0), nd["op"].param[0])
-            elif left_name == "u2":
-                left_parameters = (sympy.pi / 2, nd["op"].param[0], nd["op"].param[1])
-            elif left_name == "u3":
-                left_parameters = tuple(nd["op"].param)
-            else:
-                left_name = "u1"  # replace id with u1
-                left_parameters = (N(0), N(0), N(0))
-            # Compose gates
-            name_tuple = (left_name, right_name)
-            if name_tuple == ("u1", "u1"):
-                # u1(lambda1) * u1(lambda2) = u1(lambda1 + lambda2)
-                right_parameters = (N(0), N(0), right_parameters[2] +
-                                    left_parameters[2])
-            elif name_tuple == ("u1", "u2"):
-                # u1(lambda1) * u2(phi2, lambda2) = u2(phi2 + lambda1, lambda2)
-                right_parameters = (sympy.pi / 2, right_parameters[1] +
-                                    left_parameters[2], right_parameters[2])
-            elif name_tuple == ("u2", "u1"):
-                # u2(phi1, lambda1) * u1(lambda2) = u2(phi1, lambda1 + lambda2)
-                right_name = "u2"
-                right_parameters = (sympy.pi / 2, left_parameters[1],
-                                    right_parameters[2] + left_parameters[2])
-            elif name_tuple == ("u1", "u3"):
-                # u1(lambda1) * u3(theta2, phi2, lambda2) =
-                #     u3(theta2, phi2 + lambda1, lambda2)
-                right_parameters = (right_parameters[0], right_parameters[1] +
-                                    left_parameters[2], right_parameters[2])
-            elif name_tuple == ("u3", "u1"):
-                # u3(theta1, phi1, lambda1) * u1(lambda2) =
-                #     u3(theta1, phi1, lambda1 + lambda2)
-                right_name = "u3"
-                right_parameters = (left_parameters[0], left_parameters[1],
-                                    right_parameters[2] + left_parameters[2])
-            elif name_tuple == ("u2", "u2"):
-                # Using Ry(pi/2).Rz(2*lambda).Ry(pi/2) =
-                #    Rz(pi/2).Ry(pi-2*lambda).Rz(pi/2),
-                # u2(phi1, lambda1) * u2(phi2, lambda2) =
-                #    u3(pi - lambda1 - phi2, phi1 + pi/2, lambda2 + pi/2)
-                right_name = "u3"
-                right_parameters = (sympy.pi - left_parameters[2] -
-                                    right_parameters[1], left_parameters[1] +
-                                    sympy.pi / 2, right_parameters[2] +
-                                    sympy.pi / 2)
-            elif name_tuple[1] == "nop":
-                right_name = left_name
-                right_parameters = left_parameters
-            else:
-                # For composing u3's or u2's with u3's, use
-                # u2(phi, lambda) = u3(pi/2, phi, lambda)
-                # together with the qiskit.mapper.compose_u3 method.
-                right_name = "u3"
-                # Evaluate the symbolic expressions for efficiency
-                left_parameters = tuple(map(lambda x: x.evalf(), list(left_parameters)))
-                right_parameters = tuple(map(lambda x: x.evalf(), list(right_parameters)))
-                right_parameters = compose_u3(left_parameters[0],
-                                              left_parameters[1],
-                                              left_parameters[2],
-                                              right_parameters[0],
-                                              right_parameters[1],
-                                              right_parameters[2])
-                # Why evalf()? This program:
-                #   OPENQASM 2.0;
-                #   include "qelib1.inc";
-                #   qreg q[2];
-                #   creg c[2];
-                #   u3(0.518016983430947*pi,1.37051598592907*pi,1.36816383603222*pi) q[0];
-                #   u3(1.69867232277986*pi,0.371448347747471*pi,0.461117217930936*pi) q[0];
-                #   u3(0.294319836336836*pi,0.450325871124225*pi,1.46804720442555*pi) q[0];
-                #   measure q -> c;
-                # took >630 seconds (did not complete) to optimize without
-                # calling evalf() at all, 19 seconds to optimize calling
-                # evalf() AFTER compose_u3, and 1 second to optimize
-                # calling evalf() BEFORE compose_u3.
-            # 1. Here down, when we simplify, we add f(theta) to lambda to
-            # correct the global phase when f(theta) is 2*pi. This isn't
-            # necessary but the other steps preserve the global phase, so
-            # we continue in that manner.
-            # 2. The final step will remove Z rotations by 2*pi.
-            # 3. Note that is_zero is true only if the expression is exactly
-            # zero. If the input expressions have already been evaluated
-            # then these final simplifications will not occur.
-            # TODO After we refactor, we should have separate passes for
-            # exact and approximate rewriting.
-
-            # Y rotation is 0 mod 2*pi, so the gate is a u1
-            if (right_parameters[0] % (2 * sympy.pi)).is_zero \
-                    and right_name != "u1":
-                right_name = "u1"
-                right_parameters = (0, 0, right_parameters[1] +
-                                    right_parameters[2] +
-                                    right_parameters[0])
-            # Y rotation is pi/2 or -pi/2 mod 2*pi, so the gate is a u2
-            if right_name == "u3":
-                # theta = pi/2 + 2*k*pi
-                if ((right_parameters[0] - sympy.pi / 2) % (2 * sympy.pi)).is_zero:
-                    right_name = "u2"
-                    right_parameters = (sympy.pi / 2, right_parameters[1],
-                                        right_parameters[2] +
-                                        (right_parameters[0] - sympy.pi / 2))
-                # theta = -pi/2 + 2*k*pi
-                if ((right_parameters[0] + sympy.pi / 2) % (2 * sympy.pi)).is_zero:
-                    right_name = "u2"
-                    right_parameters = (sympy.pi / 2, right_parameters[1] +
-                                        sympy.pi, right_parameters[2] -
-                                        sympy.pi + (right_parameters[0] +
-                                                    sympy.pi / 2))
-            # u1 and lambda is 0 mod 2*pi so gate is nop (up to a global phase)
-            if right_name == "u1" and (right_parameters[2] % (2 * sympy.pi)).is_zero:
-                right_name = "nop"
-            # Simplify the symbolic parameters
-            right_parameters = tuple(map(sympy.simplify, list(right_parameters)))
-        # Replace the data of the first node in the run
-        new_op = Instruction("", [], [], [])
-        if right_name == "u1":
-            new_op = U1Gate(right_parameters[2], run_qarg)
-        if right_name == "u2":
-            new_op = U2Gate(right_parameters[1], right_parameters[2], run_qarg)
-        if right_name == "u3":
-            new_op = U3Gate(*right_parameters, run_qarg)
-
-        nx.set_node_attributes(unrolled.multi_graph, name='name',
-                               values={run[0]: right_name})
-        nx.set_node_attributes(unrolled.multi_graph, name='op',
-                               values={run[0]: new_op})
-        # Delete the other nodes in the run
-        for current_node in run[1:]:
-            unrolled._remove_op_node(current_node)
-        if right_name == "nop":
-            unrolled._remove_op_node(run[0])
-    return unrolled
 
 
 def remove_last_measurements(dag_circuit, perform_remove=True):
