+++ qiskit/transpiler/passes/__init__.py
@@ -9,6 +9,7 @@
 
 from .cx_cancellation import CXCancellation
 from .fixed_point import FixedPoint
+from .optimize_1q_gates import Optimize1qGates
 from .decompose import Decompose
 from .mapping.check_map import CheckMap
 from .mapping.basic_mapper import BasicMapper
